# Conflicts:
#	src/main/java/backend/whereIsMyTeam/config/SecurityConfig.java
#	src/main/java/backend/whereIsMyTeam/user/domain/User.java
#	src/main/java/backend/whereIsMyTeam/user/service/UserService.java

+++ src/main/java/backend/whereIsMyTeam/user/dto/UserRegisterResponseDto.java
@@ -8,12 +8,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UserRegisterResponseDto {
 
-    private Integer userIdx;
+    private Long userIdx;
 
     private String email;
 
     @Builder
-    public UserRegisterResponseDto(Integer userIdx, String email) {
+    public UserRegisterResponseDto(Long userIdx, String email) {
         this.userIdx = userIdx;
         this.email = email;
     }
